Reformat main methods
