+++ src/pages/EditorPage/Preview.js
@@ -1,12 +1,11 @@
 import React, { useCallback, useEffect, useMemo, useState } from "react";
 import { Widget } from "near-social-vm";
-import { Tab } from "./utils/const";
+import { Tab, Layout } from "./utils/const";
 import RenderPreviewButton from "./buttons/RenderPreviewButton";
 
 export default function Preview({
   tab,
   layout,
-  Layout,
   layoutClass,
   renderCode,
   jpath,
