+++ src/pages/EditorPage/TabEditor.js
@@ -1,14 +1,25 @@
 import React, { useCallback, useEffect, useMemo, useState } from "react";
 import Editor from "@monaco-editor/react";
 import { Tab } from "./utils/const";
+import prettier from "prettier";
+import parserBabel from "prettier/parser-babel";
 
-export default function TabEditor({
-  tab,
-  code,
-  widgetPath,
-  updateCode,
-  reformat,
-}) {
+export default function TabEditor({ tab, code, widgetPath, updateCode }) {
+  const reformat = useCallback(
+    (path, code) => {
+      try {
+        const formattedCode = prettier.format(code, {
+          parser: "babel",
+          plugins: [parserBabel],
+        });
+        updateCode(path, formattedCode);
+      } catch (e) {
+        console.log(e);
+      }
+    },
+    [updateCode]
+  );
+
   return (
     <div className={`${tab === Tab.Editor ? "" : "visually-hidden"}`}>
       <div
