+++ src/pages/EditorPage/Tabs.js
@@ -1,6 +1,6 @@
 import React, { useCallback, useEffect, useMemo, useState } from "react";
 
-import { Tab } from "./utils/const";
+import { Tab, Layout } from "./utils/const";
 
 export default function Tabs({
   isModule,
@@ -9,7 +9,6 @@ export default function Tabs({
   widgets,
   layout,
   setRenderCode,
-  Layout,
   code,
 }) {
   return (
